avoided some ambiguous function calls

+++ QTfrontend/hwform.cpp
@@ -310,7 +310,7 @@ void HWForm::onFrontendFullscreen(bool value)
   if (value)
     setWindowState(windowState() | Qt::WindowFullScreen);
   else {
-    setWindowState(windowState() & !Qt::WindowFullScreen);
+    setWindowState(windowState() & static_cast<int>(!Qt::WindowFullScreen));
   }
 }
